Make easier to use and find.
This makes sure the user can continue to use the api with ease (without having to update). The definition of 'cat_id2rgb' is removed from 'utils.py' and put in 'converters/panoptic2semantic_segmentation.py'.
